renamed android project
