+++ android/app/build.gradle
@@ -74,9 +74,9 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
 
-    namespace "com.ama"
+    namespace "co.askthat.qna"
     defaultConfig {
-        applicationId "co.ama"
+        applicationId "co.askthat.qna"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
